added icons for individual effects in results section
