+++ client/src/App.css
@@ -239,7 +239,7 @@ button:hover {
 }
 
 .character-image-wrapper:hover:not(.selected) {
-  background: radial-gradient(ellipse 50% 50%, var(--primary-color) 0%, rgba(0, 0, 0, 0) 120%);
+  background: radial-gradient(ellipse 50% 50%, var(--primary-color) 0%, rgba(0, 0, 0, 0) 100%);
 }
 
 .image-grid {
@@ -979,6 +979,12 @@ button:hover {
   color: #4d94ff;
 }
 
+.relic-stats-table td img {
+  width: 20px;
+  height: 20px;
+  vertical-align: middle;
+}
+
 .info-button {
   display: flex;
   background: transparent;
@@ -1368,6 +1374,7 @@ button:hover {
   padding-left: 1.2rem;
   margin: 0;
   font-size: 0.85em;
+  /* list-style-type: none; */
 }
 
 .base-relic-card li,
