+++ client/src/components/RelicResultsPage.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import RelicSlot from './RelicSlot';
 import { InformationIcon, LeftArrowIcon, RightArrowIcon, CloseIcon } from './Icons';
-import { numberFormatter } from '../utils/utils';
+import { numberFormatter, getImageUrl, getRelicDescription, getEffectIcon } from '../utils/utils';
 import relicsData from '../data/relics.json';
 
 const RelicResultsPage = ({
@@ -20,20 +20,6 @@ const RelicResultsPage = ({
   const [showTooltip, setShowTooltip] = useState(false);
   const [showRelicTooltip, setShowRelicTooltip] = useState(null);
 
-  const getImageUrl = (name, type) => {
-    if (!name) return '';
-    const cleanedName = name
-      .toLowerCase()
-      .replace(/[<>:'"/\\|?*']/g, '')
-      .replace(/ /g, '-');
-    return `/${type}/${cleanedName}.png`;
-  };
-
-  const getRelicDescription = (relicId) => {
-    if (!relicId) return null;
-    const relicEntry = relicsData[relicId.toString()];
-    return relicEntry?.description || null;
-  };
 
   const isMultipleResults = totalResults > 1;
   const showLeftArrow = isMultipleResults && currentIndex > 0;
@@ -63,7 +49,13 @@ const RelicResultsPage = ({
                 return (
                   <tr key={effectKey}>
                     <td className={isSecondary ? 'sec-effect' : ''}>
-                      • {effect.name}
+                      {isSecondary ? (
+                        <span style={{ marginLeft: '20px' }}>&nbsp;{effect.name}</span>
+                      ) : (
+                        <>
+                          <img src={getEffectIcon(effect.name)}></img> {effect.name}
+                        </>
+                      )}
                       {showScoreInfo && <span className='effect-score'>{numberFormatter.format(effect.score)}</span>}
                     </td>
                   </tr>
@@ -80,7 +72,7 @@ const RelicResultsPage = ({
                 <div className="info-tooltip">
                   <p>
                     {(() => {
-                      const description = getRelicDescription(relic.id);
+                      const description = getRelicDescription(relic.id, relicsData);
                       return description ? (
                         description.split('\n').map((line, i) => (
                           <React.Fragment key={i}>
