+++ client/src/components/RelicsPage.jsx
@@ -3,42 +3,19 @@ import items from '../data/relics.json';
 import effects from '../data/effects.json';
 import { CloseIcon, InformationIcon } from './Icons';
 import RelicFilters from './RelicFilters';
+import { EMPTY_SLOT_ID, createEffectMap, isEffectIdKnown, getEffectName } from '../utils/utils';
 
 
 
-const createEffectMap = (showDeepOfNight) => {
-  const effectMap = new Map();
-  effects.forEach(effect => {
-    // filter out deep effects if showDeepOfNight is false
-    if (effect.deep === true && !showDeepOfNight) {
-      return;
-    }
-    effect.ids.forEach(id => {
-      effectMap.set(id, effect.name);
-    });
-  });
-  return effectMap;
-};
-
-const isEffectIdKnown = (id) => {
-  return effects.some(effect => effect.ids.includes(id));
-};
 
 const RelicCard = ({ relic, items, effectMap, showRelicIdToggle }) => {
   const [showRelicId, setShowRelicId] = useState(false)
-  const EMPTY_SLOT_ID = 4294967295; // 2^32 - 1 (unsigned 32-bit integer)
 
   const relicInfo = items[relic.item_id.toString()];
 
   if (!relicInfo) {
     return null;
   }
-
-  const getEffectName = (id) => {
-    if (id === 0 || id === EMPTY_SLOT_ID) return null;
-
-    return effectMap.get(id) || `Unknown Effect (ID: ${id})`;
-  };
 
 
   const allEffects = [
@@ -52,7 +29,7 @@ const RelicCard = ({ relic, items, effectMap, showRelicIdToggle }) => {
 
   const validEffects = allEffects
     .map(effect => ({
-      name: getEffectName(effect.id),
+      name: getEffectName(effect.id, effectMap),
       isSecondary: effect.isSecondary
     }))
     .filter(effect => effect.name !== null);
@@ -111,12 +88,6 @@ const CharacterRelics = ({ characterData,
   deepRelicColorFilters,
   effectMap,
   searchTerm }) => {
-  const EMPTY_SLOT_ID = 4294967295;
-
-  const getEffectName = (id) => {
-    if (id === 0 || id === EMPTY_SLOT_ID) return null;
-    return effectMap.get(id) || `Unknown Effect (ID: ${id})`;
-  };
 
   const filteredRelics = characterData.relics.filter(relic => {
     const relicInfo = items[relic.item_id.toString()];
@@ -154,7 +125,7 @@ const CharacterRelics = ({ characterData,
       ];
 
       const effectMatches = allEffects.some(effectId => {
-        const effectName = getEffectName(effectId);
+        const effectName = getEffectName(effectId, effectMap);
         return effectName && effectName.toLowerCase().includes(searchLower);
       });
 
@@ -208,7 +179,7 @@ const RelicsPage = ({ onBack,
   const [searchTerm, setSearchTerm] = useState('');
 
   useEffect(() => {
-    setEffectMap(createEffectMap(showDeepOfNight));
+    setEffectMap(createEffectMap(showDeepOfNight, effects));
   }, [showDeepOfNight]);
 
   useEffect(() => {
@@ -220,7 +191,6 @@ const RelicsPage = ({ onBack,
           setRelicData(parsedData);
 
           const relicsWithUnknownEffects = [];
-          const EMPTY_SLOT_ID = 4294967295;
 
           parsedData.forEach(character => {
             character.relics.forEach(relic => {
@@ -229,27 +199,27 @@ const RelicsPage = ({ onBack,
                 character_name: character.character_name,
                 effect1_id: {
                   id: relic.effect1_id,
-                  known: relic.effect1_id === 0 || relic.effect1_id === EMPTY_SLOT_ID || isEffectIdKnown(relic.effect1_id)
+                  known: relic.effect1_id === 0 || relic.effect1_id === EMPTY_SLOT_ID || isEffectIdKnown(relic.effect1_id, effects)
                 },
                 effect2_id: {
                   id: relic.effect2_id,
-                  known: relic.effect2_id === 0 || relic.effect2_id === EMPTY_SLOT_ID || isEffectIdKnown(relic.effect2_id)
+                  known: relic.effect2_id === 0 || relic.effect2_id === EMPTY_SLOT_ID || isEffectIdKnown(relic.effect2_id, effects)
                 },
                 effect3_id: {
                   id: relic.effect3_id,
-                  known: relic.effect3_id === 0 || relic.effect3_id === EMPTY_SLOT_ID || isEffectIdKnown(relic.effect3_id)
+                  known: relic.effect3_id === 0 || relic.effect3_id === EMPTY_SLOT_ID || isEffectIdKnown(relic.effect3_id, effects)
                 },
                 sec_effect1_id: {
                   id: relic.sec_effect1_id,
-                  known: relic.sec_effect1_id === 0 || relic.sec_effect1_id === EMPTY_SLOT_ID || isEffectIdKnown(relic.sec_effect1_id)
+                  known: relic.sec_effect1_id === 0 || relic.sec_effect1_id === EMPTY_SLOT_ID || isEffectIdKnown(relic.sec_effect1_id, effects)
                 },
                 sec_effect2_id: {
                   id: relic.sec_effect2_id,
-                  known: relic.sec_effect2_id === 0 || relic.sec_effect2_id === EMPTY_SLOT_ID || isEffectIdKnown(relic.sec_effect2_id)
+                  known: relic.sec_effect2_id === 0 || relic.sec_effect2_id === EMPTY_SLOT_ID || isEffectIdKnown(relic.sec_effect2_id, effects)
                 },
                 sec_effect3_id: {
                   id: relic.sec_effect3_id,
-                  known: relic.sec_effect3_id === 0 || relic.sec_effect3_id === EMPTY_SLOT_ID || isEffectIdKnown(relic.sec_effect3_id)
+                  known: relic.sec_effect3_id === 0 || relic.sec_effect3_id === EMPTY_SLOT_ID || isEffectIdKnown(relic.sec_effect3_id, effects)
                 }
               };
 
@@ -283,18 +253,12 @@ const RelicsPage = ({ onBack,
       deep: { red: 0, blue: 0, yellow: 0, green: 0, purple: 0 }
     };
 
-    const EMPTY_SLOT_ID = 4294967295;
-    const getEffectName = (id) => {
-      if (id === 0 || id === EMPTY_SLOT_ID) return null;
-      return effectMap.get(id) || `Unknown Effect (ID: ${id})`;
-    };
 
     if (relicData && items) {
       relicData.forEach(character => {
         character.relics.forEach(relic => {
           const relicInfo = items[relic.item_id.toString()];
           
-          // apply same filters as CharacterRelics
           if (!showUnknownRelics && (!relicInfo || !relicInfo.name)) return;
           if (!showDeepOfNight && relicInfo && relicInfo.name && relicInfo.name.startsWith('Deep')) return;
           
@@ -314,7 +278,7 @@ const RelicsPage = ({ onBack,
             ];
 
             const effectMatches = allEffects.some(effectId => {
-              const effectName = getEffectName(effectId);
+              const effectName = getEffectName(effectId, effectMap);
               return effectName && effectName.toLowerCase().includes(searchLower);
             });
 
@@ -381,11 +345,6 @@ const RelicsPage = ({ onBack,
         if (!colorFilters[color]) return false;
       }
       if (searchTerm) {
-        const EMPTY_SLOT_ID = 4294967295;
-        const getEffectName = (id) => {
-          if (id === 0 || id === EMPTY_SLOT_ID) return null;
-          return effectMap.get(id) || `Unknown Effect (ID: ${id})`;
-        };
         const searchLower = searchTerm.toLowerCase();
         const relicName = relicInfo?.name?.toLowerCase() || '';
         const relicNameMatches = relicName.includes(searchLower);
@@ -398,7 +357,7 @@ const RelicsPage = ({ onBack,
           relic.sec_effect3_id,
         ];
         const effectMatches = allEffects.some(effectId => {
-          const effectName = getEffectName(effectId);
+          const effectName = getEffectName(effectId, effectMap);
           return effectName && effectName.toLowerCase().includes(searchLower);
         });
         if (!relicNameMatches && !effectMatches) return false;
@@ -423,11 +382,6 @@ const RelicsPage = ({ onBack,
         if (!colorFilters[color]) return false;
       }
       if (searchTerm) {
-        const EMPTY_SLOT_ID = 4294967295;
-        const getEffectName = (id) => {
-          if (id === 0 || id === EMPTY_SLOT_ID) return null;
-          return effectMap.get(id) || `Unknown Effect (ID: ${id})`;
-        };
         const searchLower = searchTerm.toLowerCase();
         const relicName = relicInfo?.name?.toLowerCase() || '';
         const relicNameMatches = relicName.includes(searchLower);
@@ -440,7 +394,7 @@ const RelicsPage = ({ onBack,
           relic.sec_effect3_id,
         ];
         const effectMatches = allEffects.some(effectId => {
-          const effectName = getEffectName(effectId);
+          const effectName = getEffectName(effectId, effectMap);
           return effectName && effectName.toLowerCase().includes(searchLower);
         });
         if (!relicNameMatches && !effectMatches) return false;
