+++ client/src/utils/hooks.js
@@ -1,0 +1,35 @@
+import { useState, useEffect } from 'react';
+
+export function usePersistentBoolean(key, defaultValue) {
+  const [value, setValue] = useState(() => {
+    const saved = localStorage.getItem(key);
+    return saved !== null ? JSON.parse(saved) : defaultValue;
+  });
+
+  useEffect(() => {
+    localStorage.setItem(key, JSON.stringify(value));
+  }, [key, value]);
+
+  return [value, setValue];
+}
+
+export function usePersistentState(key, defaultValue) {
+  const [value, setValue] = useState(() => {
+    const saved = localStorage.getItem(key);
+    if (saved === null) return defaultValue;
+    if (typeof defaultValue === 'boolean') {
+      try { return JSON.parse(saved); } catch { return saved === 'true'; }
+    }
+    return saved;
+  });
+
+  useEffect(() => {
+    if (typeof value === 'string') {
+      localStorage.setItem(key, value);
+    } else {
+      localStorage.setItem(key, JSON.stringify(value));
+    }
+  }, [key, value]);
+
+  return [value, setValue];
+}
